Renamed and updated BodyPart
