+++ game/Map.py
@@ -1,4 +1,4 @@
-from Util import *
+import Util
 
 
 class Map:
@@ -10,36 +10,17 @@ class Map:
         n_monsters = 0
         n_items = 0
         for i in range(n_monsters):
-            self.place_monsters()
-        for i in range(n_items):
-            item = random_by_level()
+            # make and insert monsters on map
             pass
-
-    def place_monsters(self):
-        pass
+        for i in range(n_items):
+            # make and insert items on map
+            pass
 
 
 class Tile:
     # The type of tile (e.g. wall, grass, floor, river)
 
-    # A list of items on the tile
-
-    # The monster on the tile
-
     # Any other data (e.g. a trap, a hole in the floor, stairs)
-
-    # For the list of items, use a linked list. This allows many items to exist
-    # on the same tile. Store a pointer to the beginning of the list.
-
-    # For the monsters, it depends. If you have more than one monster per tile,
-    # you also need a linked list. Otherwise, just keep a pointer to the
-    # monster, or some way to quickly find the monster occupying the tile.
-    # This is vital in combat.
-
-    # There's no reason, in fact, why you can't use the same linked list to
-    # store items and monsters in. It will help you keep your tiles small,
-    # and many things that can happen in the dungeon affect both monsters and
-    # items anyway so it can be handy to find them on the same list.
 
     # What other data is needed? I would use another byte to determine any
     # special dungeon features the tile carries, such as traps, stairs,
@@ -62,16 +43,14 @@ class Tile:
         self.illuminated = False
         self.glyph = ""
         self.items = []
-        self.mob = []
+        self.creature = []
         pass
 
     '''def visible_glyph(self):
-        if self.mob:
-            return self.mob.glyph
-            pass
+        if self.creature:
+            # return self.creature.glyph
         elif self.items:
             return self.items.glyph
-            pass
         else:
             return self.glyph
     '''
